Adding custom ExceptionRenderer.
This allows all error views to be contained in the Twig plugin. Plugin users will not have to worry about moving or creating these error templates into app/View/Errors.

Changed the way paths are added to the Twig FilesystemLoader class, so that invalid paths are skipped but will not cause a fatal uncaught exception. For example, app/View/plugins/MyPlugin might not exist on the filesystem, but CakePHP still considers that a valid possiblity.

Other minor cleanup to the error.css