Adding format paramter to the template reference and parser.
Going to support the format set by the RequestHandlerComponent in the filename
instead of a folder - index.xml.twig instead of xml/index.twig. Override the
View::_getViewFileName() method to try to create a parsable template string
instead of a relative path.

Starting to rename some of the error templates. The exception templates still
need some TLC to replace the default error templates and support the format
part of the file required by a TemplateReference instance.